Build: Warnings
Fixed some compiler warnings, joint ops with ebsi and stalli

review id = 10

+++ src/lib/utility/math/Vector2.h
@@ -74,7 +74,7 @@ Vector2<T>::~Vector2()
 template<class T>
 float Vector2<T>::getLengthSquared() const
 {
-	return (x * x) + (y * y);
+	return static_cast<float>(x * x) + static_cast<float>(y * y);
 }
 
 template<class T>
